tests/wpt/web-platform-tests/webmessaging/with-options/slash-origin.html

<!doctype html>
<title>special value '/'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  postMessage('', {targetOrigin: '/'});
  onmessage = this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  });
});
</script>
